*Full update to eA revision 14162. *Merged item_db, item_db2 & skill_cast_db. -Added Cooldown column. *Initialized Rune Knight skill rewrite update.
git-svn-id: https://svn.code.sf.net/p/v1-3ceam/code/trunk@29 cad27aaa-dce3-4a30-a00a-e4fd67c11881

+++ rewrite/src/map/mob.h
@@ -80,7 +80,7 @@ struct mob_db {
 	struct { int nameid,p; } mvpitem[3];
 	struct status_data status;
 	struct view_data vd;
-	short option;
+	int option; // Changed to int to give support to the new options. [pakpil]
 	int summonper[MAX_RANDOMMONSTER];
 	int maxskill;
 	struct mob_skill skill[MAX_MOBSKILL];
